aisleriot: maemo5: Fix game choice dialogue
On maemo5, the treeview in a pannable area never has any row selected,
so we failed to get the selected game type on response. Instead, the
row-activated signal is emitted when the user selects an item; so we
store the selected path in its handler and use that on response. [Found
the problem thanks to this thread on maemo-developers ML:
[http://lists.maemo.org/pipermail/maemo-developers/2009-May/019174.html
].

Bug #591847.