Fix *** glibc detected *** gtali: free(): invalid pointer.
This error happened in case where default players names was used,
in yahtzee.c "Human" string is just marked by N_() to be translated later (runtime),
but not really translated (so "Human" is assigned to DefaultPlayerNames[0]).
In players[0].name the translated string is assigned ie _("Human").
This implies that "if (players[0].name != DefaultPlayerNames[0])"
is always true because _("Human") hasn't the same address as "Human".
And we can't free a string allocated in .ro_data section.