[gnibbles] Store tail direction before reversing the worm
In the shipping code the worm-> list is reversed before the data of movement of the tail is captured. This provides the wrong answer.
The patch ensures that the direction of the tail is computed before any changes are made to the current worm data. The existing code then _correctly_ reverses the old_dir because it represents the direction of motion of the tail rather than the direction in which the tail is pointing.
The direction in which the tail is pointing prior to reversal is of course the direction that we want to finally assign as the new direction of motion.

GNOME bug #619988

Patch by Felix Oxley