Only show the preview block in the correct circumstances
This patch is closely related to 168273 and probably depends on it
http://bugzilla-attachments.gnome.org/attachment.cgi?id=168273

It stops the preview from being generated when either the preview blocks option is off, or the difficult blocks option is on.
However the preview still needs to be generated for the theme selection dialog.
So I've added a force parameter to previewBlock() function for this scenario

https://bugzilla.gnome.org/show_bug.cgi?id=617093